i got information of user and store in the localStorage

+++ src/redux/usuarioDucks.js
@@ -1,5 +1,5 @@
 // librearias
-import { auth, app } from '../firebase'
+import { auth, app, db } from '../firebase'
 
 
 // Constantes
@@ -40,20 +40,35 @@ export const ingresoUsuarioAccion = () => async (dispatch) =>{
     try {
         // const res = await axios.
         const provider  = new app.auth.GoogleAuthProvider();
-        const res = await auth.signInWithPopup(provider)
-        console.log("Esta es al respues",res);
-        dispatch({
-            type: USUARIO_EXITO,
-            payload : {
+        const res = await auth.signInWithPopup(provider)        
+        const usuario = {
+            uid: res.user.uid, 
+            email: res.user.email,
+            displayName : res.user.displayName,   
+            photoURL: res.user.photoURL
+        }
+
+        const usuarioDB = await db.collection('usuarios').doc(usuario.email).get()
+        console.log("Este es ", usuarioDB.exists);
+        if(usuarioDB.exists){
+            // cuando el usuario existe en firestore nuestros registros            
+            dispatch({
+                type: USUARIO_EXITO,
+                payload : usuarioDB.data()
+            })
+            localStorage.setItem('usuario', JSON.stringify(usuarioDB.data()))
+        }else{
+            // No existe el usuario en firestore y lo traer de google
+            await db.collection('usuarios').doc(usuario.email).set(usuario)
+            dispatch({
+                type: USUARIO_EXITO,
+                payload : usuario
+            })
+            localStorage.setItem('usuario', JSON.stringify({ 
                 uid: res.user.uid, 
                 email: res.user.email   
-            }
-        })
-        localStorage.setItem('usuario', JSON.stringify({ 
-            uid: res.user.uid, 
-            email: res.user.email   
-        }))
-        
+            }))
+        } 
     } catch (error) {
         console.log(error);
         dispatch({
